backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
pydantic-settings==2.10.1
python-multipart==0.0.20
httpx==0.28.1
pypdf==5.9.0
pytest==8.4.1
pandas==3.0.1
scikit-learn==1.8.0
numpy==2.4.3
joblib==1.5.3
kaggle==2.0.0
requests==2.32.5
Pillow==11.3.0
rapidocr-onnxruntime==1.4.4
